+++ client/src/screens/hub/HubMainScreen/components/Alert/Alert.js
@@ -1,0 +1,24 @@
+import './Alert.less';
+
+import React from 'react';
+
+import { classNames } from '../../../../../utils';
+
+function Alert({
+  children,
+}) {
+  return (
+    <div
+      className={classNames({
+        Alert: true,
+      })}
+    >
+      {children}
+    </div>
+  );
+}
+
+Alert.propTypes = {
+};
+
+export default Alert;